requirements.txt
flask
openai>=1.0.0
PyPDF2
faiss-cpu
numpy
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib